+++ src/cwtmath.h
@@ -8,11 +8,11 @@
 extern "C" {
 #endif
 
-void nsfft_exec(fft_object obj, fft_data *inp, fft_data *oup,double lb,double ub,double *w);// lb -lower bound, ub - upper bound, w - time or frequency grid (Size N)
+void nsfft_exec(fft_object obj, fft_data *inp, fft_data *oup,float lb,float ub,float *w);// lb -lower bound, ub - upper bound, w - time or frequency grid (Size N)
 
 double gamma(double x);
 
-int nint(double N);
+int nint(float N);
 
 #ifdef __cplusplus
 }
